+++ DTFX/Executors/LocalDBInsertExecutor.cs
@@ -32,7 +32,7 @@ namespace IF.Batch.DTFX.Executors
         {
             MethodBase method = MethodInfo.GetCurrentMethod();
             var element = CreateElement(rawElement);
-            TraceLog.WriteDebug(method, element.Value);
+            Logger.WriteDebug(method, element.Value);
             using (var command = new SqlCommand(element.Value))
             {
                 command.Connection = ServiceContext.GetLocalDB().Connection;
@@ -41,7 +41,7 @@ namespace IF.Batch.DTFX.Executors
                 if (!string.IsNullOrEmpty(element.ToVariable))
                 {
                     ServiceContext.SharedVariable.SetValue(element.ToVariable, result);
-                    TraceLog.WriteDebug(method, "共有変数にデータを保存しました。名前:{0}, 型:{1}", element.ToVariable, typeof(int));
+                    Logger.WriteDebug(method, "共有変数にデータを保存しました。名前:{0}, 型:{1}", element.ToVariable, typeof(int));
                 }
             }
             return ResultTypeCode.Success;