+++ DTFX/Executors/TraceLogExecutor.cs
@@ -37,7 +37,7 @@ namespace IF.Batch.DTFX.Executors
             if (!string.IsNullOrEmpty(element.ToVariable))
             {
                 ServiceContext.SharedVariable.SetValue(element.ToVariable, element.Value);
-                TraceLog.WriteDebug(method, "共有変数にデータを保存しました。名前:{0}, 型:{1}", element.ToVariable, typeof(string));
+                Logger.WriteDebug(method, "共有変数にデータを保存しました。名前:{0}, 型:{1}", element.ToVariable, typeof(string));
             }
 
             return ResultTypeCode.Success;